all good before css

+++ Steamcalculator/SatCalc.py
@@ -117,17 +117,17 @@ def SatCalcT(Input1value,Input2value,Input2):
     csat=dfsat.columns.values.tolist()
 
     if Input2=='volume':
-        vinput=int(Input2value)
+        vinput=float(Input2value)
     elif Input2=='Enthalpy':
-        hinput=int(Input2value)
+        hinput=float(Input2value)
     elif Input2=='Entropy':
-        sinput=int(Input2value)
+        sinput=float(Input2value)
     elif Input2=='Internalenergy':
-        uinput=int(Input2value)
+        uinput=float(Input2value)
     elif Input2=='Dryness':
-        xinput=int(Input2value)
+        xinput=float(Input2value)
 
-    Tinput=int(Input1value)
+    Tinput=float(Input1value)
 
     #Find closest value
     subT = dfsat['T'] - Tinput
@@ -193,4 +193,4 @@ def SatCalcT(Input1value,Input2value,Input2):
         P=DATA_VALUES[0]
         x=xinput
 
-        return[P,T,v,u,h,s,x]
+    return[P,T,v,u,h,s,x]
